[Spanify] `span` size/offset to conform to `size_t`
There's some tightening around values passed into span, with
`StrictNumeric<size_t>` being enforced. This is mostly unimportant when
dealing with compile-time values, as they can be validated on the spot,
however it does affect several places where the value is unknown at
runtime, and a signed value was being used.

Chromium changes:
https://chromium.googlesource.com/chromium/src/+/4937ad9ae64d9d3f05d13c9d041769432a12281b

```
commit 4937ad9ae64d9d3f05d13c9d041769432a12281b
Author: Peter Kasting <pkasting@chromium.org>
Date:   Tue Nov 26 22:58:16 2024 +0000

    Use StrictNumeric<size_t> for more span methods.

    Changes subspan(), split_at(), at(), and get_at() to match pre-existing
    usage in constructors, first(), last(), etc.

    Bug: 364987728
```

Resolves https://github.com/brave/brave-browser/issues/42627

+++ components/brave_wallet/common/zcash_utils.cc
@@ -53,16 +53,16 @@ std::optional<uint64_t> ReadCompactSize(base::span<const uint8_t>& data) {
   uint8_t type = data[0];
   if (data.size() > 0 && data[0] < 253) {
     value = type;
-    data = data.subspan(1);
+    data = data.subspan(1u);
   } else if (type == 253 && data.size() >= 3) {
     value = base::numerics::U16FromBigEndian(data.subspan<1, 2u>());
-    data = data.subspan(1 + 2);
+    data = data.subspan(1u + 2);
   } else if (type <= 254 && data.size() >= 5) {
     value = base::numerics::U32FromBigEndian(data.subspan<1, 4u>());
-    data = data.subspan(1 + 4);
+    data = data.subspan(1u + 4);
   } else if (data.size() >= 9) {
     value = base::numerics::U64FromBigEndian(data.subspan<1, 8u>());
-    data = data.subspan(1 + 8);
+    data = data.subspan(1u + 8);
   } else {
     return std::nullopt;
   }
@@ -97,7 +97,7 @@ std::optional<std::vector<ParsedAddress>> ParseUnifiedAddressBody(
     addr.second =
         std::vector(dejumbled_data.begin(), dejumbled_data.begin() + *size);
     result.push_back(std::move(addr));
-    dejumbled_data = dejumbled_data.subspan(*size);
+    dejumbled_data = dejumbled_data.subspan(base::checked_cast<size_t>(*size));
   }
   return result;
 }
